// frontend/src/core/enums/sqlite-worker-actions.enum.ts
export enum SQLiteWorkerActionsEnum {
	INSERT_SAMPLE = "insert_sample",
	QUERY_NOTES = "query_notes",
	SYNC_NOTES = "sync_notes",
	SYNC_NOTES_DELTA = "sync_delta",
	CLEAR_USER_DATA = "clear_user_data",
	INSERT_NOTE = "insert_note",
	UPDATE_NOTE = "update_note",
	DELETE_NOTE = "delete_note",
	GET_PENDING_NOTES = "get_pending_notes",
	CLEAR_SERVER_PENDING_SYNC = "clear_server_pending_sync",
	CLEAR_PENDING_SYNC = "clear_pending_sync",
	SYNC_NOTES_FROM_CONNECTOR = "sync_notes_from_connector",
	GET_CONNECTOR_PENDING_NOTES = "get_connector_pending_notes",
	SET_CONNECTOR_SYNCED_FOR = "set_connector_synced_for",
	GET_CONNECTOR_SYNCED_NOTES = "get_connector_synced_notes",
	GET_ALL_NOTES_FOR_EXPORT = "get_all_notes_for_export",
	REMOVE_STALE_NOTES = "remove_stale_notes",
}